Removing the connectors, and talking directly to CentralEGA
