+++ src/lega/conf/defaults.ini
@@ -5,7 +5,7 @@
 host = 127.0.0.1
 port = 9000
 
-[worker]
+[ingestion]
 # Keyserver communication
 keyserver_host = ega_keys
 keyserver_port = 9010
