+++ src/lega/verify.py
@@ -19,8 +19,10 @@ import logging
 import os
 
 from .conf import CONF
-from .utils import crypto, db, checksum
-from .utils import db_log_error_on_files, checksum
+from .utils import checksum, db_log_error_on_files
+from .utils.db import get_details
+from .utils.exceptions import VaultDecryption
+from .utils.checksum import is_valid
 from .utils.amqp import get_connection, consume
 
 LOG = logging.getLogger('verify')
@@ -30,9 +32,9 @@ def work(data):
     '''Verifying that the file in the vault does decrypt properly'''
 
     file_id = data['file_id']
-    filename, org_hash, org_hash_algo, vault_filename, vault_checksum = db.get_details(file_id)
+    filename, org_hash, org_hash_algo, vault_filename, vault_checksum = get_details(file_id)
 
-    if not checksum.is_valid(vault_filename, vault_checksum, hashAlgo='sha256'):
+    if not is_valid(vault_filename, vault_checksum, hashAlgo='sha256'):
         raise VaultDecryption(vault_filename)
 
     return { 'vault_name': vault_filename, 'org_name': filename }
